Implemented dependency inversion principle
The principle states that:
1. High-level modules should not depend on low level modules. Both should depend on the abstraction.
2. Abstractions should not depend on details. Details should depend on abstractions.

In the previous version the Player was still dependent upon the concrete implementation of Car. There is now an interface that abstracts the concrete implementation from the player. The factory is now free to return any object that implements the ICar interface. The player could recieve a GM or a Tesla and they would not need to worry about any of those details. Just that it can get a car, and that the car starts.

Second rule is covered as (ICar) does not depend on the details (Car), obviously, rather the details depend on the abstraction.

Another major advantage is that if we make the Player object accept an ICar as a parameter during construction we will be able to mock the _car object for unit testing.