Introduced simple dependency injection
It is now possible through Constructor injection to tell the Player which implementation of a Car to use. Since the Player cares so little about which whip to use, it is still possible to call the Player without providing this detail and the Player will simply call the factory for whatever Car it is currently making.