Adding Electric Car
Thanks Andrew for your pull request. I have updated the original application and its unit tests to validate that the container can in fact create a Car when I ask for a Player. I then expanded this to include an ElectricCar option, this is in a separate application. The benefit of the separate application is that it will help illustrate the power of the inversion of control container, where another team can implement their own Car and still be able to rely on the existing business logic.

I also added a small fuel | battery feature against the two car data models, I wanted to have something that separated them versus just different class names.